Move tests that run agent to it directory (IntegrationTests)
