+++ .github/workflows/scala.yml
@@ -36,7 +36,7 @@ jobs:
       with:
         java-version: 1.8
     - name: Run tests
-      run: sbt test
+      run: sbt testAll
 
   publish:
