--- otel-extension/src/test/resources/application-test.conf
@@ -1,18 +1,0 @@
-akka {
-  loglevel = DEBUG
-
-  persistence {
-    journal.plugin = "akka.persistence.journal.inmem"
-    snapshot-store.plugin = "akka.persistence.snapshot-store.local"
-    snapshot-store.local.dir = "target/snapshot"
-  }
-  actor {
-
-  }
-}
-
-io.scalac.akka-monitoring.boot.backend = false
-
-akka.stream.materializer.creation-timeout = 10 minutes
-
-
